requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
dataclasses-json==0.6.1
emoji==2.8.0
exceptiongroup==1.1.3
filetype==1.2.0
Flask==3.0.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.325
langdetect==1.0.9
langsmith==0.0.53
libmagic==1.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-iso639==2023.6.15
# python-magic-bin==0.4.14
PyYAML==6.0.1
rapidfuzz==3.4.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.27
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
